added function with five ML models
The code now includes a function that takes in a model, and the data
and it normalizes the data. It then perofrms hyper-param. tunning
for each of the five types of input models, i.e. logistic regression,
random forest, gradient boosting, svc, and knn.

It then uses the function in a for-loop structure to run each model
and finally uses the best selected model to fir the test data.